+++ pybricksdev/ble/pybricks.py
@@ -14,6 +14,9 @@ Connecting devices SHOULD then filter on this UUID to identify compatible device
 After connecting, programs SHOULD read the Software Revision characteristic of
 the Device Information Service to determine the Pybricks protocol version. This
 version can be used to determine the capabilities of the device.
+
+Additional information about the Pybricks Profile can be found at
+_`https://github.com/pybricks/technical-info/blob/master/pybricks-ble-profile.md`.
 """
 
 from enum import IntEnum, IntFlag
@@ -30,10 +33,20 @@ def _pybricks_uuid(short: int) -> str:
 PYBRICKS_SERVICE_UUID = _pybricks_uuid(0x0001)
 """The Pybricks GATT service UUID."""
 
-PYBRICKS_CONTROL_UUID = _pybricks_uuid(0x0002)
-"""The Pybricks control GATT characteristic UUID.
+PYBRICKS_COMMAND_EVENT_UUID = _pybricks_uuid(0x0002)
+"""The Pybricks command/event GATT characteristic UUID.
+
+Commands are written to this characteristic and events are received via notifications.
+
+See :class:`Command` and :class:`Event`.
 
 .. availability:: Since Pybricks protocol v1.0.0.
+"""
+
+PYBRICKS_HUB_CAPABILITIES_UUID = _pybricks_uuid(0x0003)
+"""The Pybricks hub capabilities GATT characteristic UUID.
+
+.. availability:: Since Pybricks protocol v1.2.0.
 """
 
 PYBRICKS_PROTOCOL_VERSION = semver.VersionInfo(1, 1, 0)
@@ -53,6 +66,89 @@ class Command(IntEnum):
     """Requests that the user program should be stopped.
 
     .. availability:: Since Pybricks protocol v1.0.0.
+    """
+
+    START_USER_PROGRAM = 1
+    """
+    Requests that the user program should be started.
+
+    Hub responds with :attr:`CommandError.BUSY` if a user program is currently running.
+
+    .. availability:: Since Pybricks protocol v1.2.0.
+    """
+
+    START_REPL = 2
+    """
+    Requests that the interactive REPL should be started.
+
+    Hub responds with :attr:`CommandError.INVALID_COMMAND` if the hub does not support the REPL.
+    Hub responds with :attr:`CommandError.BUSY` if a user program is currently running.
+
+    .. availability:: Since Pybricks protocol v1.2.0.
+    """
+
+    WRITE_USER_PROGRAM_META = 3
+    """
+    Requests to write user program metadata.
+
+    Parameters:
+        size: The size of the user program in bytes (32-bit unsigned integer).
+
+    Hub responds with :attr:`CommandError.BUSY` if a user program is currently running.
+
+    .. availability:: Since Pybricks protocol v1.2.0.
+    """
+
+    COMMAND_WRITE_USER_RAM = 4
+    """
+    Requests to write data to user RAM.
+
+    Parameters:
+        offset: The offset from the base user RAM address (32-bit unsigned integer).
+        payload: The data to write to this address (0 to ``max_char_size`` - 5 bytes).
+
+    Hub responds with :attr:`CommandError.VALUE_NOT_ALLOWED` if the offset and size exceeded the user RAM area.
+    Hub responds with :attr:`CommandError.BUSY` if a user program is currently running.
+
+    .. availability:: Since Pybricks protocol v1.2.0.
+    """
+
+
+class CommandError(IntEnum):
+    """
+    GATT Attribute error codes that can be sent in response to command write requests.
+    """
+
+    # NB: these values are standard BLE protocol values (i.e. Table 3.4 in v5.3 core specification)
+
+    INVALID_HANDLE = 0x01
+    """
+    The attribute handle given was not valid on this server.
+    """
+    NOT_SUPPORTED = 0x06
+    """
+    Attribute server does not support the request received from the client.
+    """
+    VALUE_NOT_ALLOWED = 0x13
+    """
+    The attribute parameter value was not allowed.
+    """
+
+    # NB: these values are limited to 0x80 – 0x9F as required by the Bluetooth
+    # spec (i.e. Table 3.4 in v5.3 core specification)
+
+    INVALID_COMMAND = 0x80
+    """
+    An invalid command was requested.
+
+    .. availability:: Since Pybricks protocol v1.2.0.
+    """
+
+    BUSY = 0x81
+    """
+    The command cannot be completed now because the required resources are busy.
+
+    .. availability:: Since Pybricks protocol v1.2.0.
     """
 
 
@@ -118,10 +214,51 @@ class StatusFlag(IntFlag):
     """
 
     SHUTDOWN = 1 << 7
-    """Hub shutdown was requested.
+    """Hub has entered shutdown state.
 
     .. availability:: Since Pybricks protocol v1.1.0.
     """
+
+    SHUTDOWN_REQUESTED = 1 << 8
+    """Hub shutdown was requested.
+
+    .. availability:: Since Pybricks protocol v1.2.0.
+    """
+
+
+class HubCapabilityFlag(IntFlag):
+    """
+    Hub capability flags.
+    """
+
+    HAS_REPL = 1 << 0
+    """
+    Indicates that the hub has an interactive REPL.
+
+    .. availability:: Since Pybricks protocol v1.2.0.
+    """
+
+    USER_PROG_MULTI_FILE_MPY6 = 1 << 1
+    """
+    Hub supports user programs using a multi-file blob with MicroPython MPY (ABI V6) files.
+
+    .. availability:: Since Pybricks protocol v1.2.0.
+    """
+
+
+def unpack_hub_capabilities(data: bytes) -> Tuple[int, HubCapabilityFlag, int]:
+    """
+    Unpacks the value read from the hub capabilities characteristic.
+
+    Args:
+        data: The raw characteristic value.
+
+    Returns:
+        A tuple of the maximum characteristic write size in bytes, the hub capability
+        flags and the max user program size in bytes.
+    """
+    max_char_size, flags, max_user_prog_size = unpack("<HII", data)
+    return max_char_size, HubCapabilityFlag(flags), max_user_prog_size
 
 
 # The Pybricks Protocol version also implies certain other services and